+++ gulpfile.js
@@ -2,7 +2,8 @@ const gulp = require('gulp');
 const sass = require('gulp-sass')(require('sass'));
 const del = require('del');
 const concat = require('gulp-concat');
-const replace = require('gulp-replace');
+const removeCode = require('gulp-remove-code');
+
 
 gulp.task('clean', function(cb) {
     del(['./dist/*']);
@@ -10,14 +11,14 @@ gulp.task('clean', function(cb) {
 })
 
 gulp.task('sass', function(cb) {
-    gulp.src('./src/**/*.scss')
+    gulp.src('./web/src/**/*.scss')
         .pipe(sass().on('error', sass.logError))
         .pipe(gulp.dest('dist'))
     cb();
 })
 
 gulp.task('copy:html', function (cb) {
-    gulp.src('./src/index.html')
+    gulp.src('./web/src/index.html')
         .pipe(gulp.dest('./dist'))
     cb()
 })
@@ -26,33 +27,32 @@ const scripts = (paths, outputFilename, outputPath) => {
     return gulp
         .src(paths)
         .pipe(concat(outputFilename))
-        // .pipe(replace(/((?!require).*)/,'//replace'))
-        .pipe(replace(/((module.exports).*)/,'//replace'))
+        .pipe(removeCode({ production: true }))
         .pipe(gulp.dest(outputPath));
 };
 
 gulp.task('copy:js', function (cb) {
-    scripts(['./src/script/**/*.js'], 'index.js', './dist/script', false);
+    scripts(['./web/src/script/**/*.js'], 'index.js', './dist/script', false);
     cb()
 })
 
 gulp.task('copy:img', function (cb) {
-    gulp.src('./src/img/**/*.{jpg,gif,png,ico}')
+    gulp.src('./web/src/img/**/*.{jpg,gif,png,ico}')
         .pipe(gulp.dest('./dist/img'))
     cb()
 })
 
 gulp.task('copy:fonts', (cb) => {
-    gulp.src('./src/fonts/**/*')
+    gulp.src('./web/src/fonts/**/*')
         .pipe(gulp.dest('./dist/fonts'))
     cb()
 })
 
 
 gulp.task('watch', () => {
-    gulp.watch('./src/**/*.scss', gulp.series('sass'));
-    gulp.watch('./src/**/*.html', gulp.series('copy:html'));
-    gulp.watch('./src/**/*.js', gulp.series('copy:js'));
+    gulp.watch('./web/src/**/*.scss', gulp.series('sass'));
+    gulp.watch('./web/src/**/*.html', gulp.series('copy:html'));
+    gulp.watch('./web/src/**/*.js', gulp.series('copy:js'));
 });
