+++ package.json
@@ -4,21 +4,24 @@
   "main": "index.js",
   "license": "MIT",
   "dependencies": {
+    "@types/jest": "^27.4.0",
     "browser-sync": "^2.27.7",
     "cors": "^2.8.5",
     "del": "^6.0.0",
     "fs-extra": "^10.0.0",
     "gulp": "^4.0.2",
-    "gulp-sass": "^5.0.0",
-    "sass": "^1.45.0nn",
     "gulp-concat": "2.6.1",
-    "gulp-replace": "1.1.3"
+    "gulp-remove-code": "^3.0.4",
+    "gulp-replace": "1.1.3",
+    "gulp-sass": "^5.0.0",
+    "jest": "^27.4.7",
+    "sass": "^1.45.0nn"
   },
   "scripts": {
     "cmd": "node_modules/.bin/gulp",
     "build": "yarn cmd default",
     "watch": "yarn cmd watch",
-    "start": "yarn cmd start"
+    "test": "jest"
   },
   "devDependencies": {
     "@types/gulp": "^4.0.9",
