+++ server/public/developers/developers.json
@@ -1,1 +1,1 @@
-[{"photo":"http://127.0.0.1:3050/photo/1642169534913developerPhoto.jpg","name":"Nikita Bezdelnikov","position":"Team Lead","height":189,"age":39,"eyeColor":"Blue","exp":14,"motherTongue":"Russian","placeOfBirth":"Horlivka","hobby":"My hobby is dancing, as it gives me joy and happiness. My hobby is dancing, as it gives me joy and happiness."},{"photo":"http://127.0.0.1:3050/photo/1642194850446developerPhoto.png","name":"Chou Itak-Soi-Det","position":"Tech Lead","height":195,"age":38,"eyeColor":"Yellow","exp":20,"motherTongue":"Russian","placeOfBirth":"Horlivka","hobby":"My hobby is dancing, as it gives me joy and happiness. My hobby is dancing, as it gives me joy and happiness."},{"photo":"http://127.0.0.1:3050/photo/1642169561337developerPhoto.jpg","name":"Rakhm Napoherov","position":"UI/UX Designer","height":165,"age":30,"eyeColor":"Green","exp":5,"motherTongue":"Russian","placeOfBirth":"Horlivka","hobby":"My hobby is dancing, as it gives me joy and happiness. My hobby is dancing, as it gives me joy and happiness."}]
+[{"photo":"http://127.0.0.1:3050/photo/1642239457247developerPhoto.jpg","name":"Nikita Bezdelnikov","position":"Team Lead","height":189,"age":39,"eyeColor":"Blue","exp":10,"motherTongue":"Russian","placeOfBirth":"Horlivka","hobby":"My hobby is dancing, as it gives me joy and happiness. My hobby is dancing, as it gives me joy and happiness."},{"photo":"http://127.0.0.1:3050/photo/1642240801160developerPhoto.jpg","name":"Chou Itak-Soi-Det","position":"Tech Lead","height":200,"age":38,"eyeColor":"Yellow","exp":20,"motherTongue":"Russian","placeOfBirth":"Horlivka","hobby":"My hobby is dancing, as it gives me joy and happiness. My hobby is dancing, as it gives me joy and happiness."},{"photo":"http://127.0.0.1:3050/photo/1642169561337developerPhoto.jpg","name":"Rakhm Napoherov","position":"UI/UX Designer","height":165,"age":30,"eyeColor":"Green","exp":5,"motherTongue":"Russian","placeOfBirth":"Horlivka","hobby":"My hobby is dancing, as it gives me joy and happiness. My hobby is dancing, as it gives me joy and happiness."}]
