+++ server/public/questions/questions.json
@@ -1,1 +1,1 @@
-{"01":[{"id":1642203042033,"question":"Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nam tristique, nunc sed lacinia feugiat, metus augue porta diam, vitae porttitor tellus magna eget tellus. Fusce a magna molestie, iaculis sapien ac, pellentesque dui. In hac habitasse platea dictum","theme":"Three theme","answer":false}]}
+{"01":[{"id":1642244334746,"question":"яфяфяф","theme":"Two theme","answer":true},{"id":1642244564489,"question":"111","theme":"Three theme","answer":true}]}
