+++ server/public/questions/questions.xml
@@ -1,1 +1,1 @@
-<?xml version="1.0"?><questions><block><id>1642203042035</id><question>Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nam tristique, nunc sed lacinia feugiat, metus augue porta diam, vitae porttitor tellus magna eget tellus. Fusce a magna molestie, iaculis sapien ac, pellentesque dui. In hac habitasse platea dictum</question><theme>Three theme</theme><answer>false</answer></block></questions>
+<?xml version="1.0"?><questions></questions>
