+++ server/public/questions/questions.yaml
@@ -1,10 +1,1 @@
- 0:
-id:1642173750873
-question:QUESTION 1
-theme:One theme
-answer:true
-1:
-id:1642203042035
-question:Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nam tristique, nunc sed lacinia feugiat, metus augue porta diam, vitae porttitor tellus magna eget tellus. Fusce a magna molestie, iaculis sapien ac, pellentesque dui. In hac habitasse platea dictum
-theme:Three theme
-answer:false
+
