+++ web/src/img/qr-code.png
