+++ web/src/index.html
@@ -118,11 +118,7 @@
                            aria-label="Add">+</a>
                         <!-- <button>+</button> -->
                     </div>
-                    <article class="text-block wrapper__text-block scrollbar" id="list-questions-add">
-                        <section class="wrapper__text-block--item"><p>dsds</p></section>
-                        <section class="wrapper__text-block--item"><p>dsds</p></section>
-                        <section class="wrapper__text-block--item"><p>dsds</p></section>
-                    </article>
+                    <article class="text-block wrapper__text-block scrollbar" id="list-questions-add"></article>
                 </main>
 
                 <div id="openModal" class="modal">
