+++ web/src/script/deleteFrom.js
