+++ web/src/style/_about.scss
