+++ web/src/style/_home.scss
