+++ web/src/style/_icons.scss
