+++ web/src/style/_variables.scss
@@ -1,7 +1,7 @@
 $main-color: #7B9E87;
 $light-main-color: #90ba9e;
 $dark-main-color: #405246;
-$sub-color: #9fbeac;
+$sub-color: #dce3dd;
 $transparent-main-color: rgba(144, 186, 158, 0.44);
 $font-family: "Roboto", sans-serif;
 $spacing: 5px;
